+++ js-node/expressjs/index.js
@@ -1,6 +1,6 @@
 (async function() {
   // const { exit } = require('process')
-  const path = require('path')
+  // const path = require('path')
   require('dotenv').config() // load
   require('dotenv').config({ path: '.env.secret', override: true }) // load
   // console.log(__dirname, process.cwd())
@@ -9,8 +9,7 @@
   // require('dotenv').config({ path: path.join(__dirname, 'apps', '.env'), override: true } )
   // require('dotenv').config({ path: path.join(__dirname, 'apps', '.env.secret'), override: true } )
   // console.log('Test Env JSON', process.env.MONGO_OPTIONS) && process.exit(0)
-
-  process.env.APP_PATH = path.join(__dirname)
+  // process.env.APP_PATH = path.join(__dirname)
 
   await require('@es-labs/node/config')( __dirname, process.cwd() )
   console.info('Globals setup and config done. Starting app... ')