+++ js-node/expressjs/router/api.js
@@ -59,6 +59,12 @@ module.exports = express.Router({caseSensitive: true})
     Promise.reject(new Error('unhandled rejection of promise')) // call on.process unhandledRejection - promise rejection, unhandled
   }))
 
+  /**
+   * GET /api/healthcheck
+   * @summary Healthcheck endpoint
+   * @tags base
+   * @return {object} 200 - success response - application/json
+   */
   .get('/healthcheck', (req, res) => res.json({ message: 'OK', app: APP_NAME, environment: process.env.NODE_ENV, version: APP_VERSION, port: API_PORT, https: Boolean(HTTPS_CERTIFICATE) }) ) // health check
 
   .post('/healthcheck', (req, res) => res.json({ message: 'POST OK' }) ) // POST health check
@@ -67,6 +73,13 @@ module.exports = express.Router({caseSensitive: true})
 
   .post('/test-cors-post', (req, res) => { res.send('Cors Done') }) // check CORS
 
+  /**
+   * GET /api/health-auth
+   * @summary Healthcheck auth endpoint
+   * @tags base
+   * @security BasicAuth
+   * @return {object} 200 - success response - application/json
+   */
   .get('/health-auth', authUser, (req, res) => { res.json({ message: 'OK' }) }) // health check auth
  
   // test uploads